@@ -1,6 +1,6 @@
 package com.alphashop.articles_web_service.exceptions;
 
-public class ItemAlreadyExistsException extends Exception {
+public class ItemAlreadyExistsException extends RuntimeException {
 
 	private static final long serialVersionUID = 5584960734228774089L;
 	private String errorMessage = "Item already exists";
