@@ -1,41 +1,30 @@
 package com.alphashop.articles_web_service.mappers;
 
-import java.util.Optional;
-
 import org.springframework.stereotype.Component;
 
-import com.alphashop.articles_web_service.common.mappers.BaseAlphaMapper;
-import com.alphashop.articles_web_service.dtos.VatDto;
+import com.alphashop.articles_web_service.dtos.VatRefDto;
+import com.alphashop.articles_web_service.dtos.VatResponseDto;
 import com.alphashop.articles_web_service.entities.Vat;
 
 @Component
-public class VatMapper extends BaseAlphaMapper<Vat, VatDto> {
+public class VatMapper {
 
-	@Override
-	public VatDto toModel(Vat entity, VatDto model) {
-		VatDto vatDto = null;
+	public VatResponseDto toModel(Vat entity) {
+		if (entity == null) return null;
 		
-		if (entity != null) {
-			vatDto = Optional.ofNullable(model).orElseGet(VatDto::new);
-			vatDto.setTaxRate(entity.getTaxRate());
-			vatDto.setDescription(entity.getDescription());
-			vatDto.setIdVat(entity.getIdVat());
-		}
+		VatResponseDto vatDto = new VatResponseDto();
+		vatDto.setTaxRate(entity.getTaxRate());
+		vatDto.setDescription(entity.getDescription());
+		vatDto.setIdVat(entity.getIdVat());
 		
 		return vatDto;
 	}
 
-	@Override
-	public Vat toEntity(VatDto model, Vat entity) {
-		Vat vat = null;
+	public Vat toEntity(VatRefDto model) {
+		if (model == null) return null;
 		
-		if (model != null) {
-			vat = Optional.ofNullable(entity).orElseGet(Vat::new);
-			vat.setTaxRate(model.getTaxRate());
-			vat.setDescription(model.getDescription());
-			vat.setIdVat(model.getIdVat());
-		}
-		
+		Vat vat = new Vat();
+		vat.setIdVat(model.getIdVat());
 		return vat;
 	}
 
