@@ -13,7 +13,7 @@ import com.alphashop.articles_web_service.entities.Article;
 
 public interface ArticleRepository extends JpaRepository<Article, String> {
 
-	Page<Article> findAllByOrderByCodArtAsc(Pageable pageRequest);
+	Page<Article> findAllByOrderByCodartAsc(Pageable pageRequest);
 	
 	// selByDescrizioneLike e findByDescrizioneLike sono alternativi.
 	// Usare findByDescriptionLike.. se si vogliono i risultati con paginazione
@@ -21,10 +21,12 @@ public interface ArticleRepository extends JpaRepository<Article, String> {
 	List<Article> selByDescriptionLike(@Param("desArt") String description);
 
 	@Query(value = "SELECT * FROM articles a WHERE UPPER(a.description) LIKE UPPER(:desArt) ORDER BY a.codart ASC", nativeQuery = true)
-	Page<Article> findByDescriptionLikeOrderByCodArtAsc(@Param("desArt") String description, Pageable pageRequest);
+	Page<Article> findByDescriptionLikeOrderByCodartAsc(@Param("desArt") String description, Pageable pageRequest);
 
-	Optional<Article> findByCodArt(String codArt);
+	Optional<Article> findByCodart(String codart);
 
 	@Query(value="SELECT a FROM Article a JOIN a.barcodes b WHERE b.barcode IN (:ean)")
 	Optional<Article> selByEan(@Param("ean") String ean);
+	
+	boolean existsByCodart(String codart);
 }
