@@ -26,7 +26,7 @@ class ArticleRestControllerDeleteIT extends BaseSpringIT {
 	
 	private String JsonData = """
 	{
-	  "codArt": "123Test",
+	  "codart": "123Test",
 	  "description": "ARTICOLO UNIT TEST INSERIMENTO",
 	  "um": "PZ",
 	  "codStat": "TESTART",
@@ -39,7 +39,7 @@ class ArticleRestControllerDeleteIT extends BaseSpringIT {
 	      "type": "CP"
 	    }],
 	  "ingredients": {
-	    "codArt": "123Test",
+	    "codart": "123Test",
 	    "info": "TEST INGREDIENTI"
 	  },
 	  "category": {
@@ -55,7 +55,7 @@ class ArticleRestControllerDeleteIT extends BaseSpringIT {
 	
 	private String JsonData02 = """
 	{
-	  "codArt": "abcTest",
+	  "codart": "abcTest",
 	  "description": "ARTICOLO UNIT TEST INSERIMENTO",
 	  "um": "PZ",
 	  "codStat": "TESTART",
@@ -68,7 +68,7 @@ class ArticleRestControllerDeleteIT extends BaseSpringIT {
 	      "type": "CP"
 	    }],
 	  "ingredients": {
-	    "codArt": "abcTest",
+	    "codart": "abcTest",
 	    "info": "TEST INGREDIENTI"
 	  },
 	  "category": {
@@ -93,7 +93,7 @@ class ArticleRestControllerDeleteIT extends BaseSpringIT {
 				.andExpect(status().isCreated())
 				.andDo(print());
 		
-		Article article = articleRepository.findByCodArt("123Test").get();
+		Article article = articleRepository.findByCodart("123Test").get();
 		assertThat(article.getDescription()).isEqualTo("ARTICOLO UNIT TEST INSERIMENTO");
 	}
 	
@@ -121,7 +121,7 @@ class ArticleRestControllerDeleteIT extends BaseSpringIT {
 				.andExpect(status().isCreated())
 				.andDo(print());
 		
-		Article article = articleRepository.findByCodArt("abcTest").get();
+		Article article = articleRepository.findByCodart("abcTest").get();
 		assertThat(article.getDescription()).isEqualTo("ARTICOLO UNIT TEST INSERIMENTO");
 	}
 		
