@@ -3,12 +3,7 @@ package com.alphashop.jwt_auth_service.security;
 import java.io.IOException;
 import java.util.Arrays;
 
-import javax.servlet.FilterChain;
-import javax.servlet.ServletException;
-import javax.servlet.http.Cookie;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-
+import org.springframework.lang.NonNull;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.core.userdetails.UserDetails;
@@ -17,6 +12,12 @@ import org.springframework.stereotype.Component;
 import org.springframework.web.filter.OncePerRequestFilter;
 
 import com.alphashop.jwt_auth_service.security.constants.PublicRoutes;
+
+import jakarta.servlet.FilterChain;
+import jakarta.servlet.ServletException;
+import jakarta.servlet.http.Cookie;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
 
 @Component
 public class JwtCookieAuthenticationFilter extends OncePerRequestFilter {
@@ -30,13 +31,13 @@ public class JwtCookieAuthenticationFilter extends OncePerRequestFilter {
     }
 
     @Override
-    protected void doFilterInternal(HttpServletRequest request, 
-    								HttpServletResponse response, 
-    								FilterChain filterChain) throws ServletException, IOException {
+    protected void doFilterInternal(@NonNull HttpServletRequest request, 
+    								@NonNull HttpServletResponse response, 
+    								@NonNull FilterChain filterChain) throws ServletException, IOException {
         
     	String path = request.getRequestURI();
 
-        // ✅ Escludi le rotte pubbliche e le richieste preflight
+        // Escludi le rotte pubbliche e le richieste preflight
         if (this.isPublicPath(path) || "OPTIONS".equalsIgnoreCase(request.getMethod())) {
 	        filterChain.doFilter(request, response);
 	        return;
