@@ -8,25 +8,63 @@ import { ArticlesComponent } from './pages/articles/articles.component';
 import { ArticleManagerComponent } from './pages/article-manager/article-manager.component';
 import { ForbiddenComponent } from './pages/forbidden/forbidden.component';
 import { Roles } from './shared/enums/roles.enum';
-import { Page50xComponent } from './pages/page-50x/page-50x.component';
 import { AuthGuard } from './core/guards/auth.guard';
 import { RoleGuard } from './core/guards/role.guard';
-import { ArticlesGridComponent } from './pages/articles-grid/articles-grid.component';
+import { ArticleDetailComponent } from './pages/article-detail/article-detail.component';
+import { Page5xxComponent } from './pages/page-5xx/page-5xx.component';
+import { PendingChangesGuard } from './core/guards/pending-changes.guard';
 
 const routes: Routes = [
-  {path:'', component:LoginComponent},
-  {path:'login', component:LoginComponent},
-  {path:'logout', component:LogoutComponent},
-  {path:'welcome', component:WelcomeComponent, canActivate:[AuthGuard, RoleGuard], data:{roles:[Roles.User]}},
-  {path:'home', component:WelcomeComponent, canActivate:[AuthGuard, RoleGuard], data:{roles:[Roles.User]}},
-  {path:'welcome/:username', component:WelcomeComponent, canActivate:[AuthGuard, RoleGuard], data:{roles:[Roles.User]}},
-  {path:'articles', component:ArticlesComponent, canActivate:[AuthGuard, RoleGuard], data:{roles:[Roles.User]}},
-  {path:'article-manager/:codArt', component:ArticleManagerComponent, canActivate:[AuthGuard, RoleGuard], data:{roles:[Roles.Admin]}},
-  {path:'article-manager', component:ArticleManagerComponent, canActivate:[AuthGuard, RoleGuard], data:{roles:[Roles.Admin]}},
-  {path:'articles/grid', component:ArticlesGridComponent, canActivate:[AuthGuard, RoleGuard], data:{roles:[Roles.User]}},
-  {path:'50x', component:Page50xComponent},
-  {path:'forbidden', component:ForbiddenComponent},
-  {path:'**', component:Page404Component}
+  { path: '', component: LoginComponent },
+  { path: 'login', component: LoginComponent },
+  { path: 'logout', component: LogoutComponent },
+  {
+    path: 'welcome',
+    component: WelcomeComponent,
+    canActivate: [AuthGuard, RoleGuard],
+    data: { roles: [Roles.User] },
+  },
+  {
+    path: 'home',
+    component: WelcomeComponent,
+    canActivate: [AuthGuard, RoleGuard],
+    data: { roles: [Roles.User] },
+  },
+  {
+    path: 'welcome/:username',
+    component: WelcomeComponent,
+    canActivate: [AuthGuard, RoleGuard],
+    data: { roles: [Roles.User] },
+  },
+  {
+    path: 'articles',
+    component: ArticlesComponent,
+    canActivate: [AuthGuard, RoleGuard],
+    data: { roles: [Roles.User] },
+  },
+  {
+    path: 'articles/detail/:codart',
+    component: ArticleDetailComponent,
+    canActivate: [RoleGuard],
+    data: { roles: [Roles.User] },
+  },
+  {
+    path: 'articles/manage/new',
+    component: ArticleManagerComponent,
+    canActivate: [AuthGuard, RoleGuard],
+    canDeactivate: [PendingChangesGuard],
+    data: { roles: [Roles.Admin] },
+  },
+  {
+    path: 'articles/manage/:codart',
+    component: ArticleManagerComponent,
+    canActivate: [AuthGuard, RoleGuard],
+    canDeactivate: [PendingChangesGuard],
+    data: { roles: [Roles.Admin] },
+  },
+  { path: '5xx', component: Page5xxComponent },
+  { path: 'forbidden', component: ForbiddenComponent },
+  { path: '**', component: Page404Component },
 ];
 
 @NgModule({
