@@ -1,380 +1,408 @@
 import { Component, OnInit} from '@angular/core';
-import { BehaviorSubject, empty, Observable } from 'rxjs';
-import { Article } from 'src/app/shared/models/Article';
+import { BehaviorSubject, catchError, combineLatest, distinctUntilChanged, EMPTY, filter, finalize, map, mergeMap, Observable, of, shareReplay, startWith, Subject, switchMap, take, takeUntil, tap, throwError } from 'rxjs';
 import { Pagination } from 'src/app/shared/models/Pagination';
 import { ArticleService } from 'src/app/shared/services/article.service';
 import { Router } from '@angular/router';
-import { HttpResponse } from '@angular/common/http';
-import { ArtStatus, ErrorMessages, StatusCodes } from 'src/app/shared/enums';
-import { ErrorResponse } from 'src/app/shared/models/ErrorResponse';
 import { LoggingService } from 'src/app/core/services/logging.service';
-import { ScrollService } from 'src/app/core/services/scroll.service';
-
-enum FilterTypes {
-  Codart = 1,
-  Description = 2,
-  Barcode = 3
-}
+import { ArticleResponse } from 'src/app/shared/models/ArticleResponse';
+import { CardFieldsConfig } from 'src/app/shared/models/CardFieldConfig';
+import { PaginatedResponseList } from 'src/app/shared/models/PaginatedResponseList';
+import { NotificationService } from 'src/app/core/services/notification.service';
+import { MESSAGE_KEYS } from 'src/app/core/i18n/message-keys';
+import { MsgKey } from 'src/app/core/i18n/msg-key.type';
+import { toMsgKey } from 'src/app/core/errors/to-msg-key.fn';
+import { toErrorViewModel } from 'src/app/core/errors/to-error-view-model.fn';
+import { TranslationService } from 'src/app/core/services/translation.service';
+import { FormBuilder, FormGroup, Validators } from '@angular/forms';
+import { LoadingStateService } from 'src/app/core/services/loading-state.service';
+import { ArticlesState, ArticlesViewModel, FilterTypes } from './ArticlesPageModel';
+import { strSanitize } from 'src/app/shared/utils/string.utils';
+import { ModalService } from 'src/app/core/services/modal.service';
 
 @Component({
   selector: 'app-articles',
   templateUrl: './articles.component.html',
-  styleUrls: ['./articles.component.scss']
+  styleUrls: ['./articles.component.scss'],
 })
 export class ArticlesComponent implements OnInit {
+  // -------------------------------------------------------------------------
+  // CONSTANTS & CONFIGURATIONS
+  // -------------------------------------------------------------------------
+  readonly NO_RESULT_MSG_KEY: MsgKey = MESSAGE_KEYS.crud.resourceNotFound;
 
-  // ENUMS
-  artStatus: typeof ArtStatus = ArtStatus;
-  errorMessages: typeof ErrorMessages = ErrorMessages;
-  statusCodes: typeof StatusCodes = StatusCodes;
+  // Configuration for input field validations
+  private readonly VALIDATION_CONFIGS: Record<string, Record<string, any>> = {
+    codart: {
+      pattern: '[a-zA-Z0-9]+',
+      minlength: 5,
+      maxlength: 20,
+    },
+    description: {
+      maxlength: 80,
+    },
+  };
 
-  serverError = false;
+  // FRONT END ERROR MESSAGES - VALIDATION
+  readonly FE_ERROR_MSGS: Record<string, Record<string, string>> = {
+    filter: {
+      required: this.translator.translate(MESSAGE_KEYS.validation.required),
+      maxlength: this.translator.translate(MESSAGE_KEYS.validation.maxLength),
+      pattern: this.translator.translate(
+        MESSAGE_KEYS.validation.lettersAndNumbers
+      ),
+    },
+  };
 
-  // Pagination
-  readonly MAX_PAG_BTNS_NR = 7;
-  readonly MINIMUM_MAX_PAG_BTNS_NR = 5;
-  paginationModel: Pagination = new Pagination(); // default values -> currentPage:1, pageSize:10
-  paginationUI: any = {};
+  // PAGINATION
+  readonly MAX_TOT_PAG_BTNS = 7;
+  private readonly MINIMUM_TOT_PAG_BTNS = 5;
 
-  // Filter
-  filter: string = '';
-  filterType: FilterTypes = FilterTypes.Description; // start with getArticlesByDesc
-  filterTypeList = [ // Used by HTML template
-    { key: "Codart", value: FilterTypes.Codart },
-    { key: "Description", value: FilterTypes.Description },
-    { key: "Barcode", value: FilterTypes.Barcode }
+  // Filter types for the template dropdown
+  readonly filterTypeList = [
+    { key: 'Codart', value: FilterTypes.Codart },
+    { key: 'Description', value: FilterTypes.Description },
+    { key: 'Barcode', value: FilterTypes.Barcode },
   ];
 
-  // They are used to block multiple requests with the same filter AND the same type.
-  lastFilter: string = this.sanitizeFilter(this.filter);
-  lastFilterType: FilterTypes = this.filterType;
+  // Mapping of search strategies based on selected filter type
+  private readonly searchStrategyMap: Record<
+    FilterTypes,
+    (
+      filter: string,
+      pagination: Pagination
+    ) => Observable<ArticleResponse | PaginatedResponseList<ArticleResponse>>
+  > = {
+    [FilterTypes.Codart]: (filter: string) =>
+      this.articleService.getArticleByCodart(filter),
+    [FilterTypes.Description]: (filter: string) =>
+      this.articleService.getArticlesByDesc(
+        filter,
+        this.stateSubject.value.pagination
+      ),
+    [FilterTypes.Barcode]: (filter: string) =>
+      this.articleService.getArticleByBarcode(filter),
+  };
 
-  errorResponse: ErrorResponse = {
-    date: new Date(),
-    status: -1,
-    message: ""
-  }
+  // Generic card component configuration
+  readonly cardFieldsConfig: CardFieldsConfig<ArticleResponse> = {
+    actionId: 'codart',
+    title: 'description',
+    text: 'codart',
+    status: 'idArtStatus',
+  };
 
-  private articlesSubject = new BehaviorSubject<Article[]>([]);
-  articles$: Observable<Article[]> = this.articlesSubject.asObservable();
 
-  codArt: string = "";
-  successMsg: string = '';
+  // -------------------------------------------------------------------------
+  // FORM & STATE MANAGEMENT
+  // -------------------------------------------------------------------------
+  private initialState: ArticlesState = {
+    articles: [],
+    pagination: new Pagination(),
+    viewMode: 'table',
+    filter: '',
+    filterType: FilterTypes.Description,
+    isMutating: false,
+    errorVM: null
+  };
+  // Every UI change or search parameter update flows through this Subject.
+  private stateSubject = new BehaviorSubject<ArticlesState>(this.initialState);
 
-  constructor(private articleService: ArticleService,
-              private router: Router,
-              private logger: LoggingService,
-              private scrollService: ScrollService) {
-    this.logger.log("constructor()");
+  // FORM GROUP
+  form!: FormGroup;
 
-    // Numero massimo di bottoni-pagina inclusi i bottoni '...' ed esclusi 'previous' e 'next'
-    // Il minimo valore impostabile è rappresentato da MINIMUM_MAX_PAG_BTNS_NR
-    // e verrà utilizzato quando 'maxPagBtnsNr' non sarà più impostato hardcoded
-    this.paginationUI.maxPagBtnsNr = this.MAX_PAG_BTNS_NR;
-    this.paginationUI.pagButtonsNr = this.paginationUI.maxPagBtnsNr;
+
+  // -------------------------------------------------------------------------
+  // REACTIVE STREAMS (DATA LOGIC)
+  // -------------------------------------------------------------------------
+
+  // Controls when to trigger a new API call based on state changes
+  private articleData$: Observable<any> = this.stateSubject.pipe(
+    // Pass to switchMap when the condition is false
+    distinctUntilChanged((prev, curr) => {
+      // 1.Check if core search parameters have changed
+      const sameSearchCriteria =
+        prev.filter === curr.filter &&
+        prev.filterType === curr.filterType &&
+        prev.pagination.currentPage === curr.pagination.currentPage;
+
+      // If search criteria are different (e.g. new page or new filter), always allow the call
+      if (!sameSearchCriteria) {
+        return false; // False = Make the call (go to switchMap)
+      }
+
+      // 2. If criteria are the same, determine if we should block the call (return true)
+
+      // CASE: Mutation Error Reset (isMutationg transition true -> false)
+      // Don't refresh the list if the mutation failed
+      const isErrorReset = prev.isMutating === true && curr.isMutating === false;
+
+      /// CASE: Stable State (isMutating transition false -> false)
+      // Don't refresh if the user clicks "Search" without changing inputs
+      const isAlreadyIdle = prev.isMutating === false && curr.isMutating === false;
+
+      return isErrorReset || isAlreadyIdle;
+    }),
+    switchMap(state => {
+      const searchObs = this.searchStrategyMap[state.filterType];
+      return searchObs(state.filter, state.pagination).pipe(
+        tap(() => {
+          if (state.isMutating) {
+            this.loader.setGlobal(false);
+            this.notificationService.setNotificationSuccess(
+              MESSAGE_KEYS.crud.deleteSuccess
+            );
+          }
+        }),
+        map(result => ({ data: result, error: null })),
+        catchError(error => {
+          this.logger.log('[Articles] Fetch data FAILED');
+          return of({ data: null, error });
+        })
+      )
+    }),
+    // Emit 'null' immediately when a new search starts.
+    // It's used by the ViewModel to show the loader before the API responds.
+    startWith(null),
+    // Share results to avoid multiple HTTP calls if multiple subscriptions (as async) exist
+    shareReplay(1)
+  );
+
+  // VIEW MODEL: The single object consumed by the template via 'async' pipe.
+  // It combines UI state, API data, and loading flags.
+  viewModel$: Observable<ArticlesViewModel> = combineLatest({
+    state: this.stateSubject.asObservable(),
+    searchResult: this.articleData$,
+    isLocalLoading: this.loader.localLoading$,
+    isGlobalLoading: this.loader.globalLoading$,
+  }).pipe(
+    tap(({searchResult}) => {
+      console.log("SEARCH_RESULT:", searchResult);
+    }),
+    map(({ state, searchResult, isLocalLoading, isGlobalLoading }) => {
+        // If 'searchResult' is null, it means startWith(null) has fired in 'articleData$'.
+        // but the server hasn't replied yet
+        const isRequestInProgress = searchResult === null;
+        const articleData = searchResult?.data;
+        const error = searchResult?.error;
+
+        // Based on the response, assign 'article list' OR 'single article'
+        const articles = articleData
+                          ? ('itemList' in articleData ? articleData.itemList : [articleData])
+                          : [];
+
+        const pagination = articleData && 'itemList' in articleData
+                            ? new Pagination(articleData.pagination)
+                            : new Pagination();
+
+        const errorFromApi = error ? toErrorViewModel(error, this.translator) : null;
+
+        // Show 'spinner' in template if the Interceptor/Service is active
+        // OR if we are waiting for a new API response ('isRequestInProgress').
+        const showLocalLoading = (isLocalLoading || isRequestInProgress) && !isGlobalLoading;
+
+        return {
+          ...state,
+          articles,
+          pagination,
+          showLoading: showLocalLoading,
+          showData: !showLocalLoading, // It also works with 404 status (resource not found).
+          showPagination: pagination.totalPages > 1 && !errorFromApi,
+          // Priority to state errors (like from deleteArt) over API fetch errors
+          errorVM: state.errorVM || errorFromApi
+        }
+      })
+    );
+
+
+  // -------------------------------------------------------------------------
+  // CONSTRUCTOR & LIFECYCLE
+  // -------------------------------------------------------------------------
+  constructor(
+    private articleService: ArticleService,
+    private router: Router,
+    private logger: LoggingService,
+    private notificationService: NotificationService,
+    private translator: TranslationService,
+    private fb: FormBuilder,
+    private loader: LoadingStateService,
+    private modalService: ModalService
+  ) {
+    // Initializing 'form'
+    this.form = this.fb.group({
+      filter: [''],
+      filterType: [FilterTypes.Description],
+    });
   }
 
   ngOnInit(): void {
-    this.logger.log("ngOnInit()");
-
-    this.getArticles();
-  }
-
-  // Rimuove spazi iniziali/finali dal filter e se è null, undefined o vuoto, restituisce stringa vuota.
-  private sanitizeFilter(filter?: string): string {
-  return filter?.trim() ?? '';
+    this.logger.log('[Articles] ngOnInit() initialized');
   }
 
 
-  setPagination = (): void => {
-    this.logger.log("setPagination()")
-    // I bottoni-pagina (pagButtons) mostrano le pagine (pages)
-    // attenzione a non confonderli
+  // -------------------------------------------------------------------------
+  // USER INTERACTION (PUBLIC METHODS)
+  // -------------------------------------------------------------------------
+  updateFilterValidators(type: FilterTypes) {
+    const control = this.form.get('filter');
+    if (!control) return;
 
-    // Se il numero totale delle pagine 'totalPages'
-    // è inferiore al numero massimo di bottoni-pagina 'maxPagBtnsNr'
-    // reimposta il numero totale di bottoni-pagina 'pagButtonsNr' che è
-    // impostato di default come il numero massimo di bottoni-pagina 'maxPagBtnsNr'
-    if (this.paginationModel.totalPages < this.paginationUI.maxPagBtnsNr) {
-      this.paginationUI.pagButtonsNr = this.paginationModel.totalPages;
+    control.clearValidators();
+
+    switch (type) {
+      case FilterTypes.Description:
+        control.setValidators([
+          Validators.maxLength(
+            this.VALIDATION_CONFIGS['description']['maxlength']
+          ),
+        ]);
+        break;
+      case FilterTypes.Codart:
+        control.setValidators([
+          Validators.required,
+          Validators.pattern(this.VALIDATION_CONFIGS['codart']['pattern']),
+          Validators.maxLength(this.VALIDATION_CONFIGS['codart']['maxlength']),
+        ]);
+        break;
+      case FilterTypes.Barcode:
+        control.setValidators([Validators.required]);
+        break;
     }
-    this.logger.log("paginationUI: ", this.paginationUI);
+
+    control.updateValueAndValidity();
   }
 
-  refresh = (): void => {
-    this.logger.log("refresh()");
+  onSubmit = (): void => {
+    this.logger.log('[Articles] refresh() triggered');
 
-    const sanitizedFilter = this.sanitizeFilter(this.filter);
+    // If the form is invalid, mark as touched, clear the ViewModel, clear last search markers and stop
+    if (this.form.get('filter')?.invalid) {
+      this.logger.warn('[Articles] refresh() -> Form invalid, aborting search');
 
-    if (sanitizedFilter === this.lastFilter && this.filterType === this.lastFilterType) {
-      this.logger.log("Refresh skipped: same filter and type.");
+      this.form.get('filter')?.markAsTouched();
       return;
     }
 
-    this.logger.log("Refresh executed.");
+    const formFilter = strSanitize(this.form.get('filter')?.value);
+    const formFilterType = this.form.get('filterType')?.value;
 
-    this.paginationModel = new Pagination();
-    this.resetResponses();
-    this.getArticles();
-
-    this.lastFilter = sanitizedFilter;
-    this.lastFilterType = this.filterType;
-  }
-
-  private getLiteralArtStatus = (idArtStatus: string): string => {
-    const mapArtStatus: Record<string, ArtStatus> = {
-      '1': ArtStatus.Active,
-      '2': ArtStatus.Suspended,
-      '3': ArtStatus.Deleted
-    }
-    return mapArtStatus[idArtStatus] ?? ArtStatus.Error;
-  }
-
-  // ***************************************
-  // ******** CRUD Articles - Start ********
-  // ***************************************
-  resetResponses = (): void => {
-    this.successMsg = '';
-    this.errorResponse = {
-      date: new Date(),
-      status: -1,
-      message: ""
-    }
-  }
-
-  // READ - CRUD Articles
-  getArticles = (): void => {
-    this.logger.log("getArticles()");
-
-    const observable: Observable<HttpResponse<any>> = this.getObservableByFilterType();
-    this.logger.log("observable -> ", observable);
-
-    observable.subscribe({
-      next: response => this.handleSuccessGET(response),
-      error: error => this.handleErrorGET(error)
+    // Triggering a search simply means updating the State.
+    this.updateState({
+      filter: formFilter,
+      filterType: formFilterType,
+      pagination: new Pagination(),
+      isMutating: false,
+      errorVM: null
     });
+
+    this.logger.log('Refresh executed.');
+  };
+
+  goToCreate() {
+    this.logger.log('[Articles] Redirect to create page');
+    this.router.navigate(['articles/manage/new']);
   }
-  private getObservableByFilterType(): Observable<HttpResponse<any>> {
-    this.logger.log("getObservableByFilterType()");
-    this.logger.log("filter -> " + this.filter);
-    this.logger.log("filterType -> " + this.filterType);
-
-    const sanitizedFilter = this.sanitizeFilter(this.filter);
-
-    const filterMap: Record<FilterTypes, () => Observable<HttpResponse<any>>> = {
-      [FilterTypes.Codart]: () => this.articleService.getArticleByCodart(sanitizedFilter),
-      [FilterTypes.Description]: () => this.articleService.getArticlesByDesc(sanitizedFilter, this.paginationModel),
-      [FilterTypes.Barcode]: () => this.articleService.getArticleByBarcode(sanitizedFilter)
-    }
-
-    const getObs: () => Observable<HttpResponse<any>> = filterMap[this.filterType];
-
-    if (!getObs) {
-      throw new Error("Invalid filterType");
-    }
-
-    return getObs();
+  goToEdit(codart: string) {
+    this.logger.log('[Articles] Redirect to edit page');
+    this.router.navigate(['articles/manage', codart]);
   }
-  private handleSuccessGET = (response: any): void => {
-    this.logger.log("---- H A N D L E  R E S P O N S E ----");
-    this.logger.log("handleResponse()");
-    this.logger.log(response);
+  goToDetail(codart: string) {
+    this.logger.log('[Articles] Redirect to detail page');
+    this.router.navigate(['articles/detail', codart]);
+  }
 
-    // In base alla risposta, assegna la lista di articles OR il singolo article
-    const articles: Article[] = response.body.itemList || [response.body];
+  onDelete(codart: string) {
+    this.logger.log(`[Articles] Delete requested for: ${codart}`);
 
-    // Converte lo stato id numerico di ogni article in letterale (es. 1 -> 'Attivo')
-    articles.map(art => {
-      if (art.idArtStatus) {
-        art.idArtStatus = this.getLiteralArtStatus(art.idArtStatus);
+    this.modalService.open({
+      title: 'Confirm Deletion',
+      message: `Are you sure you want to permanently delete article <b>${codart}</b>?`,
+      confirmText: 'Delete Article',
+      type: 'danger'
+    }).subscribe(confirmed => {
+      this.logger.log(`[Articles] Delete confirmation result: ${confirmed}`);
+      if (confirmed) {
+        this.deleteArt(codart);
       }
+    })
+  }
+
+  // Invoked on click of a page button of the 'Pagination' component
+  pageChange = (currentPage: number): void => {
+    this.logger.log(`[Articles] pageChange() -> Moving to page ${currentPage}`);
+
+    this.updateState({
+      pagination: new Pagination({
+        ...this.stateSubject.value.pagination,
+        currentPage
+      }),
     });
+  };
 
-    // Setta la paginazione se la risposta la fornisce
-    if (response.body.pagination) {
-      this.paginationModel = response.body.pagination;
-      this.setPagination();
+  // Switch view mode between Table and Grid
+  switchView() {
+    const viewMode = this.stateSubject.value.viewMode == 'table' ? 'grid' : 'table'
+    this.logger.log(`[Articles] switchView() -> ${viewMode}`);
+    this.updateState({viewMode});
+  }
+
+
+  // -------------------------------------------------------------------------
+  // TEMPLATE HELPERS
+  // -------------------------------------------------------------------------
+
+  // USED BY TEMPLATE: Assign a CSS class to form fields
+  // based on their validity and state (dirty/touched).
+  getValidationClass(field: string): string {
+    const formControl = this.form.get(field);
+
+    if (
+      !formControl ||
+      (!formControl.dirty && !formControl.touched) ||
+      formControl.disabled
+    ) {
+      return '';
     }
-
-    this.articlesSubject.next(articles);
-  }
-  // Gestione errore per GET (lettura articoli)
-  private handleErrorGET = (error: any): void => {
-    // Puliamo la lista degli articoli in caso di errore
-    this.articlesSubject.next([]);
-
-    const sanitizedFilter = this.sanitizeFilter(this.filter);
-
-    const messages: Record<FilterTypes, { empty: string; filled: string }> = {
-      [FilterTypes.Description]:{
-        empty: `Currently there are no articles.`,
-        filled: `We couldn't find any articles matching '${sanitizedFilter}'.`
-      },
-      [FilterTypes.Codart]:{
-        empty: `Insert a Codart.`,
-        filled: `The article with Codart '${sanitizedFilter}' is not found.`
-      },
-      [FilterTypes.Barcode]:{
-        empty: `Insert a Barcode.`,
-        filled: `The article with Barcode '${sanitizedFilter}' is not found.`
-      }
-    };
-
-    // Messaggio personalizzato per NotFound (404)
-    const isEmpty = sanitizedFilter === '';
-    const notFoundMsg =
-      isEmpty
-        ? messages[this.filterType].empty
-        : messages[this.filterType].filled;
-
-
-    // Passiamo l'errore e il messaggio personalizzato per NotFound (404)
-    this.processError(error, notFoundMsg);
+    return formControl.valid ? 'is-valid' : 'is-invalid';
   }
 
-  // CREATE - CRUD Articles
-  createArt() {
-    this.router.navigate(['article-manager']);
-  }
 
-  // UPDATE - CRUD Articles
-  updateArt(codArt: string) {
-    this.router.navigate(['article-manager', codArt]);
-  }
+  // -------------------------------------------------------------------------
+  // PRIVATE LOGIC AND API OPERATIONS
+  // -------------------------------------------------------------------------
+  private deleteArt = (codart: string): void => {
+    this.logger.log(`[Articles] deleteArt() -> Target: ${codart}`);
 
-  // DELETE - CRUD Articles
-  deleteArt = (codArt: string): void => {
-    this.resetResponses();
-    this.codArt = codArt;
+    this.loader.setGlobal(true);
+    let isSuccess = false;
 
-    this.articleService.deleteArticleByCodart(codArt).subscribe({
-      next: this.handleSuccessDELETE,
-      error: this.handleErrorDELETE
-      // error: this.handleErrorResp
+    this.articleService.deleteArticleByCodart(codart).pipe()
+      .subscribe({
+        next: () => {
+          this.logger.log(`[Articles] deleteArt() -> Success: ${codart} deleted`);
+          isSuccess = true;
+          this.updateState({isMutating: true, errorVM: null});
+        },
+        error: (error) => {
+          this.logger.error(`[Articles] deleteArt() -> Error deleting ${codart}`, error);
+
+          this.loader.setGlobal(false);
+
+          const errorVM = toErrorViewModel(error, this.translator);
+
+          this.updateState({isMutating: false, errorVM});
+
+          if (error.status !== 403) {
+            this.notificationService.setNotificationError(toMsgKey(errorVM.code!));
+          }
+        }
+      });
+  };
+
+  // UPDATE COMPONENT STATE
+  private updateState(partialState: Partial<ArticlesState>): void {
+    this.stateSubject.next({
+      ...this.stateSubject.value,
+      ...partialState,
     });
   }
-  private handleSuccessDELETE = (response: any): void => {
-    this.logger.log("handleSuccessResp()");
-    this.logger.log(response);
-
-    this.successMsg = response.body.message;
-
-    this.filterType = FilterTypes.Description;
-    this.getArticles();
-
-    //  Scroll down the page to the alert element with the response message
-    this.scrollService.scrollToAnchor('successAlert');
-  }
-  private handleErrorDELETE = (error: any): void => {
-    // Per la cancellazione possiamo usare sempre lo stesso messaggio se non trovato
-    this.processError(error, ErrorMessages.ElementNotFound);
-  }
-
-  // Metodo generale per processare un errore HTTP.
-  // Gestisce sia gli errori con JSON inviato dal backend sia errori senza JSON.
-  private processError(error: any, notFoundMsg: string): void {
-    this.logger.log("processError()", error);
-
-    // error.error -> risposta JSON inviata dal backend di tipo ErrorResponse
-    //               (es. { status: 404, message: "Article not found" })
-    //
-    // error.status -> il codice HTTP reale ricevuto dal protocollo (es. 404, 500, 403)
-
-    // Se il backend invia JSON, prendiamo il codice dal JSON (backend status)
-    // Altrimenti fallback sul codice HTTP reale (protocollo)
-    this.errorResponse.status = error.error?.status ?? error.status ?? -1;
-
-    // Se il backend invia JSON, prendiamo il messaggio dal JSON (backend message)
-    // altrimenti fallback su un messaggio generico.
-    this.errorResponse.message = error.error?.message || ErrorMessages.GenericError;
-
-    // Se l’errore è un NotFound imposta il messaggio personalizzato che è stato passato
-    if (error.status === StatusCodes.NotFound) {
-      this.errorResponse.message = notFoundMsg;
-    }
-
-    // Scrolla la pagina verso l’alert degli errori
-    this.scrollService.scrollToAnchor('errorAlert');
-  }
-  // *************************************
-  // ******** End - CRUD Articles ********
-  // *************************************
-
-  // ******** paginationUI - Start ********
-  // pageNr: numeroPagina passato dalla pagina HTML al metodo
-  // this.paginationUI.pagButtonsNr: Attuale numero di bottoni-pagina visualizzato
-  // this.paginationModel.currentPage: numeroPagina corrente
-  // this.paginationModel.totalPages: totale delle pagine messe a disposizione dal server
-
-  // Invocato al click del singolo bottone-pagina
-  pageChange = (currentPage: string|number): void => {
-    this.paginationModel.currentPage = typeof(currentPage)==='string' ? parseInt(currentPage) : currentPage;
-
-    // Reimposta 'filterType' con FilterTypes.ByDesc così
-    // 'this.getArticles()' invoca 'getObservableByFilterType()'
-    // che a sua volta invoca 'getArticlesByDesc()'.
-    // Questo è l'unico dei 3 metodi GET che invia al server
-    // i valori di 'paginationModel' tramite url
-    // (es. http://localhost:5051/api/articles/cerca/descrizione/pane?currentPage=1&pageSize=10)
-    this.filterType = FilterTypes.Description;
-
-    this.resetResponses();
-    this.getArticles();
-  }
-
-  // Ritorna un booleano per mostrare o meno il 'pageNr' su uno dei bottoni-pagina
-  shouldDisplayPageNr(pageNr: number): boolean {
-    const firstPage = 1;
-    const lastPage = this.paginationModel.totalPages;
-    const currentPage = this.paginationModel.currentPage;
-    const maxPagBtnsNr = this.paginationUI.maxPagBtnsNr; // 7
-    const constNrBtns = 2; // Il primo e l'ultimo bottone-pagina hanno valore costante
-    const isFirstOrLastPage = (pageNr === firstPage || pageNr === lastPage);
-
-    return (
-            isFirstOrLastPage
-            ||
-            (pageNr >= 1 && pageNr <= firstPage+(maxPagBtnsNr-constNrBtns))
-              &&
-            (currentPage >= firstPage && currentPage <= 4)
-            ||
-            (pageNr >= lastPage-(maxPagBtnsNr-constNrBtns) && pageNr <= lastPage)
-              &&
-            (currentPage >= lastPage-3 && currentPage <= lastPage)
-            ||
-            (currentPage >= firstPage+4 && currentPage <= lastPage-4)
-              &&
-            (pageNr >= currentPage-2 && pageNr <= currentPage+2)
-           );
-  }
-
-  isCurrentPageNr(pageNr: number): boolean {
-    return this.paginationModel.currentPage === pageNr;
-  }
-
-  // Ritorna il valore da stampare sul singolo bottone-pagina
-  displayPageNr(pageNr: number): string {
-    const firstPage = 1;
-    const lastPage = this.paginationModel.totalPages;
-    const currentPage = this.paginationModel.currentPage;
-    const maxPagBtnsNr = this.paginationUI.maxPagBtnsNr; // 7
-    const constNrBtns = 2; // Il primo e l'ultimo bottone-pagina hanno valore costante
-
-    const showEllipsis = (
-                          (currentPage >= firstPage && currentPage <= 4)
-                            &&
-                          pageNr == firstPage+(maxPagBtnsNr-constNrBtns)
-                          ||
-                          (currentPage >= lastPage-3 && currentPage <= lastPage)
-                            &&
-                          pageNr == lastPage-(maxPagBtnsNr-constNrBtns)
-                          ||
-                          (currentPage >= firstPage+4 && currentPage <= lastPage-4)
-                            &&
-                          (pageNr == currentPage-2 || pageNr == currentPage+2)
-                        );
-
-    return showEllipsis ? "..." : pageNr.toString();
-  }
-// ******** End - Pagination ********
 }
