@@ -1,23 +1,11 @@
-import { Component, OnInit } from '@angular/core';
-import { Observable } from 'rxjs';
-import { LoadingService } from 'src/app/core/services/loading.service';
-import { LoggingService } from 'src/app/core/services/logging.service';
+import { Component, Input } from '@angular/core';
 
 @Component({
-  selector: 'app-spinner',
+  selector: 'app-spinner[isVisible]', // TODO in Angular 16+: Remove [isVisible] and migrate to @Input({required: true})
   templateUrl: './spinner.component.html',
   styleUrls: ['./spinner.component.scss']
 })
-export class SpinnerComponent implements OnInit {
-
-  loading$!: Observable<boolean>;
-
-  constructor(private loader: LoadingService, private logger: LoggingService) {
-    this.logger.log('SpinnerComponent constructor');
-  }
-
-  ngOnInit(): void {
-    this.loading$ = this.loader.loading$;
-    this.loading$.subscribe(value => this.logger.log('SpinnerComponent loading$ -> ', value));
-  }
+export class SpinnerComponent {
+  @Input() isVisible!: boolean;
+  @Input() isGlobal: boolean = false;
 }
