@@ -1,27 +1,52 @@
 import { Component, Input } from '@angular/core';
-import { NgModel } from '@angular/forms';
-import { log } from 'console';
+import { AbstractControl, FormControl } from '@angular/forms';
 
 @Component({
   selector: 'app-validation-messages',
-  templateUrl: './validation-messages.component.html'
+  templateUrl: './validation-messages.component.html',
 })
 export class ValidationMessagesComponent {
-  @Input() errors: string[] | null = null;
-  @Input() ngModel!: NgModel;
-  @Input() defaultMsg: string = 'Required field**';
-  @Input() customMsgs?: { [key: string]: string };
+  @Input() formCtrl?: AbstractControl | null;
+  @Input() beErrorMsgs: string[] = [];
+  @Input() feErrorMsgs?: Record<string, string>;
 
-  hasNgValidation(): boolean {
+  readonly defaultErrorMsg: string = 'Validation error';
 
+  readonly defaultErrorMsgs: Record<string, string> = {
+    required: 'This field is required',
+    pattern: 'Invalid format',
+    minlength: 'Too short',
+    maxlength: 'Too long',
+    min: 'Value too small',
+    max: 'Value too large',
+  };
 
-    var res = Boolean(
-      this.ngModel &&
-      this.ngModel.invalid &&
-      (this.ngModel.dirty || this.ngModel.touched)
-    )
-    console.log('hasNgValidation', this.ngModel, res, this.errors);
+  readonly IS_VALID_MSG = 'Looks good!';
 
-    return res;
+  showValidMsg() {
+    return (
+      !!this.formCtrl &&
+      (this.formCtrl.dirty || this.formCtrl.touched) &&
+      this.formCtrl.valid &&
+      this.beErrorMsgs.length === 0
+    );
+  }
+
+  isFrontendInvalid(): boolean {
+    return (
+      !!this.formCtrl &&
+      (this.formCtrl.dirty || this.formCtrl.touched) &&
+      this.formCtrl.invalid
+    );
+  }
+
+  getFeErrorMsg() {
+    const errorKey = Object.keys(this.formCtrl?.errors ?? {})[0];
+
+    return (
+      this.feErrorMsgs?.[errorKey] ||
+      this.defaultErrorMsgs[errorKey] ||
+      this.defaultErrorMsg
+    );
   }
 }
