@@ -1,6 +1,7 @@
-export enum ArtStatus {
+export enum ItemStatus {
   Active = 'Active',
   Suspended = 'Suspended',
   Deleted = 'Deleted',
-  Error = 'Error'
+  Error = 'Error',
+  None = 'None'
 };
