@@ -1,8 +1,10 @@
-export enum ErrorMessages {
-  UnavailableServer = 'The server is temporarily unavailable',
-  OperationNotAllowed = "Operation not allowed",
-  GenericError = 'An error occurred',
-  ElementNotFound = 'Element not found',
-  AuthenticationException = "Authentication Failed",
-  ValidationError = "Validation Error"
-};
+// export enum ErrorMessages {
+//   UnavailableServer = 'The service is temporarily unavailable, please try again shortly.',
+//   OperationNotAllowed = "Operation not allowed",
+//   GenericError = 'An error occurred, please try again shortly.',
+//   ElementNotFound = 'Element not found',
+//   AuthenticationException = "Authentication Failed",
+//   ValidationError = "Validation Error",
+//   ItemAlreadyExists = "An item with this identifier already exists",
+//   NoChangesDetected = "No changes detected."
+// };
