@@ -1,4 +1,4 @@
 export interface Barcode {
   barcode: string;
-  type: string;
+  idTypeArt: string;
  }
