@@ -1,4 +1,4 @@
 export interface Category {
-  id: number;
-  description: string;
+  id: number,
+  description: string
  }
