@@ -1,8 +1,8 @@
-import { ErrorValidationMap } from "src/app/shared/types/ErrorValidationMap";
+import { ErrorValidationMap } from "src/app/shared/types/ErrorValidationMap.type";
 
 export interface ErrorResponse {
-  date: Date;
-	status: number;
-	message: string;
-	errorValidationMap?: ErrorValidationMap;
+  date: Date,
+	status: number,
+	code: string,
+	errorValidationMap?: ErrorValidationMap | null
 }
